Move SQL parser and planner into SQL module
