+++ rust/datafusion/src/sql/mod.rs
@@ -15,17 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-//! DataFusion is a modern distributed compute platform implemented in Rust that uses
-//! Apache Arrow as the memory model
+//! SQL parser and query planner
 
-extern crate arrow;
-#[macro_use]
-extern crate serde_derive;
-extern crate serde_json;
-extern crate sqlparser;
-
-pub mod datasource;
-pub mod execution;
-pub mod logicalplan;
-pub mod optimizer;
-pub mod sql;
+pub mod parser;
+pub mod planner;
